pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "puya-odoo-mcp"
version = "0.1.0"
description = "MCP server for Odoo 16 with RBAC"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["src"]
